+++ src/main/java/ch/bfh/ti/i4mi/mag/Config.java
@@ -198,6 +198,12 @@ public class Config {
     @Value("${mag.ch-epr-fhir.epr-spid-as-patientid:false}")
     private boolean chEprspidAsPatientId;
 
+    @Value("${mag.ch-epr-fhir.ch-mhd:false}")
+    private boolean chMhd;
+
+    @Value("${mag.ch-epr-fhir.ch-ppqm:false}")
+    private boolean chPpqm;
+
 
     /**
      * baseurl of gateway