+++ docs/changelog.md
@@ -1,4 +1,4 @@
-## 2024/05/15 v068
+## 2024/05/15 v070
 - support for multiple IDP's [128](https://github.com/i4mi/MobileAccessGateway/issues/128)
 - reenable $find-medication-list
 - use IdP also for patient access in web app