+++ docs/changelog.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Fixed an NPE in the assertion route
+
 ## 2024/05/15 v070
 - support for multiple IDP's [128](https://github.com/i4mi/MobileAccessGateway/issues/128)
 - reenable $find-medication-list