+++ docs/changelog.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Fixed an NPE in the assertion route
+- capability statment validation issues [177](https://github.com/i4mi/MobileAccessGateway/issues/177)
 
 ## 2024/05/15 v070
 - support for multiple IDP's [128](https://github.com/i4mi/MobileAccessGateway/issues/128)