+++ docs/changelog.md
@@ -2,7 +2,9 @@
 
 - Fixed an NPE in the assertion route
 - Improved mapping between XDS XON and FHIR Organization [165](https://github.com/i4mi/MobileAccessGateway/issues/165)
-- capability statment validation issues [177](https://github.com/i4mi/MobileAccessGateway/issues/177)
+- capability statement validation issues [#177](https://github.com/i4mi/MobileAccessGateway/issues/177)
+- Add CH PIXm ITI-83 constraints for sourceIdentifier and targetSystem [#170](https://github.com/i4mi/MobileAccessGateway/issues/170)
+
 
 ## 2024/05/15 v070
 - support for multiple IDP's [128](https://github.com/i4mi/MobileAccessGateway/issues/128)